refactor(discriminators): move EMCResolvers mapping to resolvers/index
